Rework the layout to be more general purpose
This is to adapt the layout to more versatile stuff, instead of just
being a resume page.
